add DbManager#getTable
