+++ xutils/src/main/java/org/xutils/db/table/TableEntity.java
@@ -18,15 +18,12 @@ package org.xutils.db.table;
 import android.database.Cursor;
 
 import org.xutils.DbManager;
-import org.xutils.common.util.DoubleKeyValueMap;
 import org.xutils.common.util.IOUtil;
 import org.xutils.db.annotation.Table;
 import org.xutils.ex.DbException;
 
 import java.lang.reflect.Constructor;
 import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 
 
 public final class TableEntity<T> {
@@ -37,19 +34,14 @@ public final class TableEntity<T> {
     private ColumnEntity id;
     private Class<T> entityType;
     private Constructor<T> constructor;
+    private volatile boolean checkedDatabase;
 
     /**
      * key: columnName
      */
     private final LinkedHashMap<String, ColumnEntity> columnMap;
 
-    /**
-     * key: dbName#className
-     */
-    private static final DoubleKeyValueMap<DbManager, Class<?>, TableEntity<?>> tableMap
-            = new DoubleKeyValueMap<DbManager, Class<?>, TableEntity<?>>();
-
-    private TableEntity(DbManager db, Class<T> entityType) throws Throwable {
+    /*package*/ TableEntity(DbManager db, Class<T> entityType) throws Throwable {
         this.db = db;
         this.entityType = entityType;
         this.constructor = entityType.getConstructor();
@@ -69,50 +61,6 @@ public final class TableEntity<T> {
 
     public T createEntity() throws Throwable {
         return this.constructor.newInstance();
-    }
-
-    @SuppressWarnings("unchecked")
-    public static <T> TableEntity<T> get(DbManager db, Class<T> entityType) throws DbException {
-        synchronized (tableMap) {
-            TableEntity<T> table = (TableEntity<T>) tableMap.get(db, entityType);
-            if (table == null) {
-                try {
-                    table = new TableEntity<T>(db, entityType);
-                } catch (Throwable ex) {
-                    throw new DbException(ex);
-                }
-                tableMap.put(db, entityType, table);
-            }
-
-            return table;
-        }
-    }
-
-    public static void remove(DbManager db, Class<?> entityType) {
-        synchronized (tableMap) {
-            tableMap.remove(db, entityType);
-        }
-    }
-
-    public static void remove(DbManager db, String tableName) {
-        synchronized (tableMap) {
-            if (tableMap.size() > 0) {
-                Class<?> key2 = null;
-                ConcurrentHashMap<Class<?>, TableEntity<?>> cls_table_map = tableMap.get(db);
-                for (Map.Entry<Class<?>, TableEntity<?>> entry : cls_table_map.entrySet()) {
-                    TableEntity table = entry.getValue();
-                    if (table != null) {
-                        if (table.getName().equals(tableName)) {
-                            key2 = entry.getKey();
-                            break;
-                        }
-                    }
-                }
-                if (key2 != null) {
-                    tableMap.remove(db, key2);
-                }
-            }
-        }
     }
 
     public boolean tableIsExist() throws DbException {
@@ -164,13 +112,11 @@ public final class TableEntity<T> {
         return columnMap;
     }
 
-    private volatile boolean checkedDatabase;
-
-    public boolean isCheckedDatabase() {
+    /*package*/ boolean isCheckedDatabase() {
         return checkedDatabase;
     }
 
-    public void setCheckedDatabase(boolean checkedDatabase) {
+    /*package*/ void setCheckedDatabase(boolean checkedDatabase) {
         this.checkedDatabase = checkedDatabase;
     }
